+++ examples/classic_controllers/custom_classic_controllers_ind_motor_example.py
@@ -18,7 +18,7 @@ if __name__ == '__main__':
 
     motor_type = 'SCIM'
     control_type = 'SC'
-    action_type = 'AbcCont'
+    action_type = 'Cont'
 
     env_id = action_type + '-' + control_type + '-' + motor_type + '-v0'