+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+## Added
+- State Action processors as a new feature to process actions and states directly in the environment.
 
 ## [1.0.1] - 2021-12-20
 ## Added